Rework help a bit.
