+++ packages/cli/src/helpers/mapCommandMetadata.ts
@@ -1,0 +1,11 @@
+import { CommandMetadataMap, Commandable } from '../types';
+
+export default function mapCommandMetadata(commands: { [path: string]: Commandable }) {
+  const map: CommandMetadataMap = {};
+
+  Object.entries(commands).forEach(([path, config]) => {
+    map[path] = config.getMetadata();
+  });
+
+  return map;
+}
